requirements.txt
einops==0.6.1
einops-exts==0.0.3
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.17.1
open3d==0.17.0
opencv-python==4.10.0.84
Pillow==9.3.0
PyYAML==6.0
fvcore==0.1.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.13.1
smplx==0.1.28
tqdm==4.64.1
traitlets==5.14.1
transformers==4.28.1
transforms3d==0.4.1
wandb==0.13.5
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0